+++ ReleaseNotes.md
@@ -1,1 +1,2 @@
-* Expose the DiscretesAsStrings option on AnalyticQuery
+* Expose the "next maintenance window" in the EMS system info.
+* Include support for the Admin User GetUsers route.